cherez/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0
django-ckeditor==6.7.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-widget-tweaks==1.5.0
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
Pillow==10.1.0
polib==1.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
